requirements.txt
# Libraries required for HuggingFace and LangChain
# pip install -q -U -r requirements.txt

huggingface_hub
transformers
accelerate 
bitsandbytes
langchain
langchain-huggingface
langchain-core 
langchain_community
bs4 
requests 
faiss-cpu
langchain-ollama 
langchain-google-genai
langchain-openai
streamlit
pypdf
sentence-transformers
openai
sentencepiece 
google-generativeai


